% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{add_alliance}
\alias{add_alliance}
\title{Add column variants, including summaries, and deflated dollars, to a contract dataset.}
\usage{
add_alliance(
  df,
  isoAlpha3_col = "ISOalpha3",
  drop_col = FALSE,
  prefix = NULL,
  purge_unprefixed = FALSE,
  skip_name = FALSE,
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/"
)
}
\arguments{
\item{df}{A data frame.}

\item{path}{The path or url for the column key.  By default, checks
the CSISdefense Github lookups repository at CSISdefense/Lookup/master/}

\item{isoAlpha3_col="ISOalpha3"}{The column with country codes in 3 letter ISO format}

\item{drop_col=FALSE}{If true drop many of the supporting columns, by default FALSE.}

\item{prefix=NULL}{If provided, add to the start of all the added variables (e.g. Place, Vendor, Origin). Default NULL.}

\item{purge_unprefixed=FALSE}{If TRUE, drop all columns previously added by this function that lack a prefix. Used to avoid adding the same columns repeatedly.}

\item{skip_name=FALSE}{If true, do not add the country name field}
}
\value{
The fiscal year of the date.
}
\description{
Add column variants, including summaries, and deflated dollars, to a contract dataset.
}
\details{
Add column variants, including summaries, and deflated dollars,
to a contract dataset.
}
