% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{all_duplicate}
\alias{all_duplicate}
\title{All Duplicates}
\usage{
all_duplicate(x, key = NULL)
}
\arguments{
\item{x}{the data frame to be checked}

\item{key}{list of one or more column names that are suspected to be the key}
}
\value{
All instances of rows where the primary key is duplicated
}
\description{
All Duplicates
}
\details{
Duplicated just returns the 2nd row of duplicates and doesn't have
a trivially easy way of just checking for duplicates in primary keys rather
than all rows, this function covers both.
}
