% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{build_plot}
\alias{build_plot}
\title{Adds a geom layer to a ggplot object based on user input.}
\usage{
build_plot(
  data,
  chart_geom = "Line Chart",
  share = FALSE,
  x_var = NULL,
  y_var = "None",
  color_var = "None",
  facet_var = "None",
  second_var = NULL,
  legend = TRUE,
  caption = TRUE,
  labels_and_colors = NULL,
  column_key = NULL,
  format = FALSE,
  ytextposition = FALSE,
  first_color_on_bottom = TRUE,
  alpha_var = NULL,
  invert_bool = NULL,
  suppress_x_var_newline = TRUE,
  hide_unlabeled = FALSE
)
}
\arguments{
\item{data}{A tibble of formatted data for the ggplot}

\item{chart_geom}{choose from Line Chart or Bar Chart; defaults to Line Chart}

\item{share}{If TRUE, calculates the share}

\item{x_var}{The name of fiscal year variable, as string}

\item{y_var}{The name of variable to plot on y-axis}

\item{color_var}{The name of the coloration variable, as string}

\item{facet_var}{The name of facet variable, as string}

\item{second_var}{The name of the secondary facet variable, as string}

\item{legend}{If TRUE, includes a legend}

\item{caption}{If TRUE, includes a source caption}

\item{labels_and_colors}{A csis360 lookup data.frame with factor information}

\item{column_key}{A csis360 lookup data.frame with column information}

\item{format}{If TRUE, summarize the data.frame}

\item{ytextposition}{If TRUE, add ytextposition to allow for geom_text overlays.}

\item{alpha_var}{Variable for setting the transparency of bars or line type of lines, coded to be used with year to ate.}

\item{invert_bool}{Used to create population pyramid or import/export charts, specifies when to show data in the negative space of the y-axis.}

\item{suppress_x_var_newline}{For categorical x-axis variables, remove any new line from the character string}

\item{hide_unlabeled}{If true, check that unlabeled variables are 1% of the date or less and if so hide them.}
}
\value{
A ggplot object including user-specified geom layer
}
\description{
Adds a geom layer to a ggplot object based on user input.
}
\details{
Intended to handle ggplot settings that depend on user input.
Settings that apply universally should be added in server.R.
}
