% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{check_derived}
\alias{check_derived}
\title{Check Derived}
\usage{
check_derived(x, key, derived_col, na.rm = FALSE)
}
\arguments{
\item{x}{the data frame to be checked}

\item{key}{list of one or more column names that are suspected to be the key}

\item{derived_col}{check whether this variable only varies with the primary key}

\item{na.rm}{whether to ignore na derived columns}
}
\value{
True if the derived_col varies only with primary key.
}
\description{
Check Derived
}
\details{
A derived column is one that could be consolidated only to the primary keys
and the derived column
}
