% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(x, key)
}
\arguments{
\item{x}{the data frame to be checked}

\item{key}{list of one or more column names that are suspected to be the key}
}
\value{
True if the columns are unique identifiers, false with warning otherwise.
}
\description{
Check Key
}
\details{
This function is uesed to check if a provided set of columns act as
unique identifiers. This is often particularly valuable before merging two
sets of tables, though it has a variety of other uses.
}
\examples{

}
