% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{clear_edit_ui}
\alias{clear_edit_ui}
\title{Removes the variable and value selection selectInputs from the Edit Data tab}
\usage{
clear_edit_ui(input, session = getDefaultReactiveDomain())
}
\arguments{
\item{input}{Shiny input object}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\description{
Removes the variable and value selection selectInputs from the Edit Data tab
}
