% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{create_edit_values_list}
\alias{create_edit_values_list}
\title{Creates the list of values available for editing, when the user changes the variable they are examining}
\usage{
create_edit_values_list(current_data, input,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{current_data}{The current data frame for the app}

\item{input}{Shiny input object}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\value{
Returns???
}
\description{
Creates the list of values available for editing, when the user changes the variable they are examining
}
