% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{date_x_year_breaks}
\alias{date_x_year_breaks}
\title{Quickly assign yearly breaks to a chart}
\usage{
date_x_year_breaks(
  start,
  stop,
  by,
  fiscal_year = TRUE,
  partial_year = NULL,
  partial_label = "\\nYTD"
)
}
\arguments{
\item{start}{First year of break sequence}

\item{stop}{Last year of break sequence}

\item{by}{Frequency of data breaks, e.g. 1 for every year, 5 for every 5 years}

\item{fiscal_year}{A placeholder for future tuning by fiscal vs. calendar year}

\item{partial_year}{If one year of incomplete data is included, specify it with this variable}

\item{partial_label}{If one year of incomplete data is included, specify it with this variable}
}
\value{
A plot with added color and fill scales for the column passed
}
\description{
Quickly assign yearly breaks to a chart
}
\details{
Add year breaks at specified intervals for date data
}
\examples{
date_x_year_breaks(2000,2023,2)

}
