% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{deflate}
\alias{deflate}
\title{Deflation using GitHub-based CSV file}
\usage{
deflate(
  data,
  money_var = "Amount",
  fy_var = "Fiscal_Year",
  deflator_file = "Lookup_Deflators.csv",
  deflator_var = "OMB25_GDP23",
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/",
  directory = "economic/",
  deflator_dropped = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{money_var}{The quoted name of the dollar-value variable}

\item{fy_var}{The quoted name of the fiscal year variable}

\item{deflator_file}{The quoted file name of the deflators to use;
must be a CSV with the column "Fiscal_Year."}

\item{deflator_var}{The quoted name of the defalator variable variable,
by default "deflator.2017"}

\item{path}{The path or url for the deflator_file CSV.  By default, checks
the CSISdefense Github lookups repository at CSISdefense/csis360/Lookup-Tables/data/economic/}
}
\value{
Returns a data frame with the money_var deflated, using deflator_var
otherwise identical to the original data frame
}
\description{
Deflation using GitHub-based CSV file
}
\details{
Warning: This function should be used __in data processing only__,
not in a live app.  It reads an external file from GitHub,
which will slow down an app substantially if done repeatedly.
}
\examples{

FullData <- read_csv("2017_SP_CompetitionVendorSizeHistoryBucketPlatformSubCustomer.csv",
  col_names = TRUE, col_types = "cccccccccc",na=c("NA","NULL"))
PrepareLabelsAndColors(Coloration,FullData,"Customer")

}
