% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{drop_from_frame}
\alias{drop_from_frame}
\title{Filters out and drops factor levels from a factor in a data frame}
\usage{
drop_from_frame(
  passed_frame,
  passed_var,
  passed_levels,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{passed_frame}{A data frame, as an object}

\item{passed_var}{A variable, as a string}

\item{passed_levels}{The levels to drop, as a string}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\value{
The data frame with the factor level removed
}
\description{
Filters out and drops factor levels from a factor in a data frame
}
