% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{factor_wrap}
\alias{factor_wrap}
\title{Add line breaks to a string in a factor}
\usage{
factor_wrap(f, nwrap = 20)
}
\arguments{
\item{nwrap}{The number of characters between returns}

\item{string}{A string}
}
\value{
String with carriage breaks
}
\description{
Add line breaks to a string in a factor
}
