% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{FactorToNumber}
\alias{FactorToNumber}
\title{Turn a factor  into a number}
\usage{
FactorToNumber(x)
}
\arguments{
\item{x}{The factor or character string}
}
\value{
The same column in numerical format.
}
\description{
Turn a factor  into a number
}
\details{
Returns the variable after removing $s and ,s.
}
\examples{
FactorToNumber("5")
}
