% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{fill_derived}
\alias{fill_derived}
\title{Fill Derived}
\usage{
fill_derived(x, key, derived_col)
}
\arguments{
\item{x}{the data frame to be checked}

\item{key}{list of one or more column names that are suspected to be the key}

\item{derived_col}{check whether this variable only varies with the primary key}
}
\value{
x with any na values filled in, if all non-na values are consistent.
}
\description{
Fill Derived
}
\details{
A derived column is one that could be consolidated only to the primary keys
and the derived column (with any nas removed). If that criteria is met, this function
then fills in the nas with those derived alues.
}
\examples{

}
