% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{format_period_average}
\alias{format_period_average}
\title{Returns data in the appropriate format for the user-specified plot}
\usage{
format_period_average(data, period_var, y_var, breakout, labels_and_colors)
}
\arguments{
\item{data}{}

\item{period_var}{The variable with the period designations, one per entry}

\item{y_var}{}

\item{breakout}{Facet and/or color}

\item{labels_and_colors}{}
}
\value{

}
\description{
Returns data in the appropriate format for the user-specified plot
}
