% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{get_base_folder}
\alias{get_base_folder}
\title{Get Base Folder}
\usage{
get_base_folder(folder)
}
\arguments{
\item{folder}{The folder to check}
}
\value{
"../../[folder]", "../[folder]", or "[folder]" depending on the relative position.
}
\description{
Get Base Folder
}
\details{
When a project is open, R files use the base of the repository
as the working directory. RMD files use their own position, typically
script or analysis.
}
