% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{get_delim}
\alias{get_delim}
\title{Return the appropriate delimeter for the file type}
\usage{
get_delim(filename)
}
\arguments{
\item{filename}{The name of the data file}
}
\value{
The delimter to use.
}
\description{
Return the appropriate delimeter for the file type
}
\details{
Returns ',' for csv and '/t' for txt files. Creates an error for other file types.
}
\examples{
get_delim("test.csv")
}
