% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{get_fiscal_year}
\alias{get_fiscal_year}
\title{Get Column Key based on the names in a data frame}
\usage{
get_fiscal_year(x)
}
\arguments{
\item{x}{A date.}
}
\value{
The fiscal year of the date.
}
\description{
Get Column Key based on the names in a data frame
}
\details{
Converts the date to fiscal year by
extracting the date and adding 1 if the month
is october or later.
}
\examples{

get_fiscal_year(as.Date('2019-12-10'))

}
