% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{get_label}
\alias{get_label}
\title{Get Label}
\usage{
get_label(var, column_key, share = FALSE)
}
\arguments{
\item{var}{The names of the column, default of none for generic colors}

\item{column_key}{A csis360 lookup data.frame with column information}

\item{share}{If TRUE, calculates the share????????? COPIED FROM STAND}
}
\value{
character string with the label corresponding to a column
}
\description{
Get Label
}
