% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{get_legend}
\alias{get_legend}
\title{Extract a legend}
\usage{
get_legend(a.gplot)
}
\arguments{
\item{a.gplot}{?????????????}
}
\value{
Returns a legend (from ________________)
}
\description{
Extract a legend
}
