% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{get_local_lookup_path}
\alias{get_local_lookup_path}
\title{A kludge, tests a variety of local paths to find the lookup-tables reposistory}
\usage{
get_local_lookup_path()
}
\value{
The path, if a known one exists. Otherwise it will throw an error.
}
\description{
A kludge, tests a variety of local paths to find the lookup-tables reposistory
}
\details{
The lookup-table repository is a public one and can be found online.
However, when updating lookup files, it can be faster to use the local version
rather than wait for github to full parse the pushed files. In addition a
local path can enable working offline.
}
\examples{
sget_local_lookup_path()

}
