% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{get_local_sharepoint_path}
\alias{get_local_sharepoint_path}
\title{A kludge, that should someday be replaced by using Microsoft365 repo.}
\usage{
get_local_sharepoint_path(site = "DIIG - Documents")
}
\value{
The path, if a known one exists. Otherwise it will throw an error.
}
\description{
' @param site Which site. The DIIG sharepoint folder is the default.
}
\details{
This is strictly for internal CSIS usage, as the user will need to
separately be syncing sharepoint to their drive. The intent is to update,
particularly via log_path,
}
\examples{
get_local_sharepoint_path()

}
