% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{ggsave600dpi}
\alias{ggsave600dpi}
\title{Extract a legend https://stackoverflow.com/questions/43366616/ggplot2-legend-only-in-a-plot
Alternate unused approach From https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
\usage{
ggsave600dpi(
  filename,
  gg,
  width,
  height,
  units = "in",
  size = 12,
  lineheight = 0.13,
  caption_fraction = 10/12,
  ...
)
}
\arguments{
\item{filename}{Output file}

\item{gg}{A ggplot}

\item{width}{Output height as per ggsave}

\item{height}{Output height as per ggsave}

\item{units}{Output measurement unit as per ggsave}

\item{size}{New text size}

\item{caption_fraction}{Ratio of caption to the rest of text}
}
\value{
Does not return, outptus to file
}
\description{
Extract a legend https://stackoverflow.com/questions/43366616/ggplot2-legend-only-in-a-plot
Alternate unused approach From https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
