% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{group_by_list}
\alias{group_by_list}
\title{Group By List}
\usage{
group_by_list(x, key)
}
\arguments{
\item{x}{the data frame to be checked}

\item{key}{list of one or more column names that are suspected to be the key}
}
\value{
Group_By using a list of quoted names.
}
\description{
Group By List
}
\details{
Replacement for group_by_ now that it has been depricated.
}
\examples{

}
