% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{import_SQLserver_file}
\alias{import_SQLserver_file}
\title{Import a comparison file from SQL server.}
\usage{
import_SQLserver_file(Path, prefix, file.name)
}
\arguments{
\item{Path}{File path}

\item{prefix}{prefix to add to file name, e.g. defense}

\item{file.name}{file name, sans any prefix.}
}
\value{

}
\description{
Import a comparison file from SQL server.
}
\details{
Quality control function
}
