% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{jitter_binary}
\alias{jitter_binary}
\title{jitter_binary}
\usage{
jitter_binary(a, jitt = 0.05)
}
\value{
Takes a binary variable and randomly spreads it within 0 and 1
}
\description{
jitter_binary
}
\details{
From Andrew Gelman and Jennifer Hill. For use in graphs
to show the quantity of variable that otherwise would be stuck
on a single lines at 0 and 1. This is different from geom_jitter
because it keeps all the data within the possible range of the variable,
not allowing values greater than 1 or less than 0.
}
\examples{
jitter_binary(contract$b_CBre)

}
