% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{label_top}
\alias{label_top}
\title{Label Top X entries in a data frame}
\usage{
label_top(
  df,
  col,
  n = 5,
  weight = "Action_Obligation_OMB24_GDP22",
  top_name = NA,
  group_list = NA,
  time = "Fiscal_Year",
  recent = NA,
  retain_rank = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{col}{The column to be evaluated.}

\item{weight}{The variable summed to determine top entries, by default Action_Obligation_OMB24_GDP22}

\item{top_name}{The new column that will include}

\item{time}{The variable used for when considering recent top entries, by default Fiscal_Year}

\item{recent}{Value used to give extra weight to recently top entries rather than just the overall period}

\item{retain_rank}{Retain any ranked variables created in processing.}

\item{group}{Sub groupings by which to rank, by default not included}
}
\value{
The revised data frame with the new variable
}
\description{
Label Top X entries in a data frame
}
\details{
Add column variants, including summaries, and deflated dollars,
to a contract dataset.
}
