% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{label_units}
\alias{label_units}
\title{Automatically put an axis in a standardized unit.}
\usage{
label_units(x)
}
\arguments{
\item{x}{A column of data}
}
\value{
Data in a standardized form.
}
\description{
Automatically put an axis in a standardized unit.
}
