% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{LatticePlotWrapper_csis360}
\alias{LatticePlotWrapper_csis360}
\title{A legacy function for graphs that allows primary and s econdary facets and}
\usage{
LatticePlotWrapper_csis360(VAR.color.legend.label, VAR.main.label,
  VAR.X.label, VAR.Y.label, VAR.Coloration, VAR.long.DF, VAR.ncol = NA,
  VAR.x.variable, VAR.y.variable, VAR.y.series, VAR.facet.primary = NA,
  VAR.facet.secondary = NA, MovingAverage = 1, MovingSides = 1,
  DataLabels = NA, caption = FALSE, legend = FALSE,
  geom.text.size = 2.5)
}
\arguments{
\item{VAR.color.legend.label}{Label for the fill legend}

\item{VAR.main.label}{Main title for the graph}

\item{VAR.X.label}{X-axis label}

\item{VAR.Y.label}{Y-axis label}

\item{VAR.Coloration}{A csis360 lookup data.frame with factor information}

\item{VAR.long.DF}{A data frame of formatted data for the ggplot}

\item{VAR.x.variable}{The name of fiscal year variable, as string}

\item{VAR.y.variable}{The name of variable to plot on y-axis}

\item{VAR.y.series}{The name of the coloration variable, as string}

\item{VAR.facet.primary}{The name of the primary facet variable, as string. Defaults to NA.}

\item{VAR.facet.secondary}{The name of the secondary facet variable, as string. Defaults to NA.}

\item{MovingAverage}{Hww many units to include in a moving average, defaults to 1, i.e. no average.}

\item{MovingSides}{How to center the moving average.}

\item{DataLabels}{Whether to draw data labels on the graph, default=NA or no.}

\item{caption}{If TRUE, includes a source caption, default is false.}

\item{legend}{If TRUE, includes a legend, default is false.}

\item{geom.text.size}{Size multipllier for text on the graph.}

\item{ncol}{Number of columns for a facet_wrap, doesn't apply if there's a secondary facet}

\item{share}{If TRUE, calculates the share}
}
\value{
A ggplot object following user-specified parameters.
}
\description{
A legacy function for graphs that allows primary and s econdary facets and
}
\details{
This should really be integrated into build_plot
}
\examples{

}
