% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{LimitData}
\alias{LimitData}
\title{Limit the data to only comparison set e.g. only services or defense.}
\usage{
LimitData(Path, FPDS.gov, customer = NULL, big.ProdServ = NULL)
}
\arguments{
\item{Path}{File path}

\item{FPDS.gov}{dataframe}

\item{customer}{character string; choose from "Defense", "DHS", "State and IAP", or ""; defaults to "" for all}

\item{big.ProdServ}{character string; choose from "Products", "Services", or ""; defaults to "" for all}
}
\value{
updated dataframe
}
\description{
Limit the data to only comparison set e.g. only services or defense.
}
\details{
Quality control function
}
