% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{LimitScope}
\alias{LimitScope}
\title{Match up yeaers between two datasets to limit the scope.}
\usage{
LimitScope(FPDS.gov, main.DF)
}
\arguments{
\item{FPDS.gov}{FPDS.gov data frame}

\item{main.DF}{Comparison  data frame.}
}
\description{
Match up yeaers between two datasets to limit the scope.
}
\details{
Quality control function
}
