% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{load.FPDS.gov.buckets.df}
\alias{load.FPDS.gov.buckets.df}
\title{Load product or service code downloads from FPDS.gov}
\usage{
load.FPDS.gov.buckets.df(
  Path = "https://github.com/CSISdefense/Lookup-Tables",
  prod_serv_blank = "",
  customer = ""
)
}
\arguments{
\item{Path}{file path name where Lookup files are stored}

\item{prod_serv_blank}{character string; choose from "Products", "Services", or ""; defaults to "" for all}

\item{customer}{character string; choose from "Defense", "DHS", "State and IAP", or ""; defaults to "" for all}
}
\value{
dataframe
}
\description{
Load product or service code downloads from FPDS.gov
}
\details{
Quality control function
}
