% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{load.FPDS.gov.customers.df}
\alias{load.FPDS.gov.customers.df}
\title{Returns a dataframe with customer data for quality control check}
\usage{
load.FPDS.gov.customers.df(Path, Choice.Data, Which.Data = "")
}
\arguments{
\item{Path}{file path name where Footing files are stored}

\item{Which.Data}{character string determining which data files to call; defaults to ""}
}
\value{
dataframe
}
\description{
Returns a dataframe with customer data for quality control check
}
\details{
Quality control function
}
