% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{log_plot}
\alias{log_plot}
\title{Save a copy of the plot, a current dollars csv, and an excel copy}
\usage{
log_plot(
  plot,
  df,
  filename,
  xlsx,
  sheet,
  path = "..\\\\output",
  width = 6.5,
  height = 3.5,
  startRow = 1,
  startCol = 10,
  format = TRUE,
  var_list = NA,
  output_plot = TRUE,
  x_var = NULL,
  y_var = NULL
)
}
\arguments{
\item{plot}{a ggplot object}

\item{df}{the underlying data}

\item{filename}{the name for the files, excluding extension}

\item{xlsx}{the excel file to output to}

\item{sheet}{the sheet to use in excel, typically shorter than the name}

\item{path="..\\output\\"}{what directory for the output}

\item{startCol=10}{}
}
\value{
no value
}
\description{
Save a copy of the plot, a current dollars csv, and an excel copy
}
