% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{lookup_colors}
\alias{lookup_colors}
\title{Create a palette for ggplot2 use, using lookup using GitHub-based CSV file}
\usage{
lookup_colors(data_for_lookup, category_vars = c(all_factors(),
  all_characters()), lookup_file = "lookup_coloration_table.csv",
  path = "https://raw.githubusercontent.com/CSISdefense/hamre_lookups/master/")
}
\arguments{
\item{data_for_lookup}{A data frame}

\item{category_vars}{A vector containing the quoted names of the variables
for which to set colors.  Any category variable that may be used as
a breakout should be included.  By default, includes all factor or character
variables.}

\item{lookup_file}{The quoted file name of the color lookup to use;
must be a CSV with the columns "variable" and "hexvalue".}

\item{path}{The path or url for the data_for_lookup CSV.  By default, checks
the CSISdefense Github lookups repository at CSISdefense/hamre_lookups/master}
}
\value{
Returns a named vector of hex values, suitable for use with ggplot
functions such as scale_color_manual() and scale_fill_manual()
}
\description{
Create a palette for ggplot2 use, using lookup using GitHub-based CSV file
}
\section{Warning}{
 This function should be ideally be used in data
processing, to create a colorset vector that will be saved in a .Rda file
for app use.  If you want to use it directly in an app, make sure it
only runs once - over the full data set - at app startup.  It reads an
external (GitHub) file and is not optimized for speed.
}

\examples{
RDTE_data <- deflate(
  data_to_deflate = RDTE_data,
  money_var = "Millions",
  fy_var = "fiscal_year")

}
