% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLookups.R
\name{NA.check}
\alias{NA.check}
\title{An internal function to check for NAs in columns after a join}
\usage{
NA.check(VAR.df, VAR.input, VAR.output, VAR.file)
}
\arguments{
\item{VAR.df}{A data frame}

\item{VAR.input}{The column(s) that had been used to join}

\item{VAR.output}{The column(s) that result from the join}

\item{VAR.file}{The file used in the join}
}
\value{
None. Instead the function raises an error if there are NAs.
}
\description{
An internal function to check for NAs in columns after a join
}
\examples{
NA.check(VAR.existing.df,
  VAR.input=by,
  VAR.output=NA.check.columns,
  VAR.file=VAR.file)

}
