% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{na_check}
\alias{na_check}
\title{An internal function to check for NAs in columns after a join}
\usage{
na_check(data, input_var, output_var, lookup_file)
}
\arguments{
\item{data}{A data frame}

\item{input_var}{The var(s) that had been used to join}

\item{output_var}{The var(s) that result from the join}

\item{lookup_file}{The file used in the join}
}
\value{
None. Instead the function raises an error if there are NAs
}
\description{
An internal function to check for NAs in columns after a join
}
\details{
This function is intended to catch gaps in lookup tables
and to alert the developer before they can come into use. The core intent
is to throw an error message that, if needed, will guide the developer to
the file they need to update and the rows they need to add.
}
\examples{
na_check(data,
  input_var=by,
  output_var=add_var,
  lookup_file=lookup_file)

}
