% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{populate_edit_var}
\alias{populate_edit_var}
\title{Populates the edit_var element on the edit page, based on the current data}
\usage{
populate_edit_var(current_data, input, session = getDefaultReactiveDomain())
}
\arguments{
\item{current_data}{The current data frame for the app}

\item{input}{Shiny input object}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\description{
Populates the edit_var element on the edit page, based on the current data
}
