% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{populate_ui_var_lists}
\alias{populate_ui_var_lists}
\title{Fills the ui menus with appropriate variables from the tibble passed to it}
\usage{
populate_ui_var_lists(data_source, session = getDefaultReactiveDomain())
}
\arguments{
\item{data_source}{A tibble from which to populate the ui menus}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\description{
Fills the ui menus with appropriate variables from the tibble passed to it
}
