% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{prepare_labels_and_colors}
\alias{prepare_labels_and_colors}
\title{Prepare Labels And Colors}
\usage{
prepare_labels_and_colors(data, var = NULL, na_replaced = FALSE,
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/style/",
  missing_allowed = FALSE)
}
\arguments{
\item{data}{The data frame to be joined}

\item{var}{Prepare colors for this columns}

\item{na_replaced}{If true, replace NAs for var before adding colors}

\item{path}{The location of the lookup file}
}
\value{
A new data frame for build on the var var, it will
include colors, and order, and proper name labels.
}
\description{
Prepare Labels And Colors
}
\details{
This function applies standard colors and orders to a single
data frame var. Colors and order are drawn from pre-existing lookup tables.
When values are missing or wrong, these tables must be manually updated.
This function is badly optimized, reading in multiple csvs every time.
It is intend for use in data preparation source code and not to be used in a
real time web environment.
}
\examples{
FullData<-standardize_variable_names(Path,
  FullData)

}
