% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{PrepareLabelsAndColors}
\alias{PrepareLabelsAndColors}
\title{Deflation using GitHub-based CSV file}
\usage{
PrepareLabelsAndColors(VAR.Coloration, VAR.long.DF, VAR.y.series,
  ReplaceNAs = FALSE)
}
\arguments{
\item{data_to_deflate}{A data frame}

\item{money_var}{The quoted name of the dollar-value variable}

\item{fy_var}{The quoted name of the fiscal year variable}

\item{deflator_file}{The quoted file name of the deflators to use;
must be a CSV with the columns "FY" and "Deflator."}

\item{path}{The path or url for the deflator_file CSV.  By default, checks
the CSISdefense Github lookups repository at CSISdefense/hamre_lookups/master}
}
\value{
Returns a data frame with the money_var deflated, otherwise identical
to the original data frame
}
\description{
Deflation using GitHub-based CSV file
}
\section{Warning}{
 This function should be used __in data processing only__,
not in a live app.  It reads an external file from GitHub,
which will slow down an app substantially if done repeatedly.
}

\examples{
RDTE_data <- deflate(
  data_to_deflate = RDTE_data,
  money_var = "Millions",
  fy_var = "fiscal_year")

}
