% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{read.tables}
\alias{read.tables}
\title{Bind multiple csv files into one dataframe}
\usage{
read.tables(file.names, ...)
}
\arguments{
\item{file.names}{list each filename of data with same column names}
}
\value{
combined data from multiple files into one data table
with an extra column identifying which file the line of data came from
}
\description{
Bind multiple csv files into one dataframe
}
\examples{
data <- read.tables(c("filename1.csv", "filename2.csv"))


}
