% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLookups.R
\name{read_and_join}
\alias{read_and_join}
\title{Read in an external file and join it with an existing data frame.}
\usage{
read_and_join(VAR.path, VAR.file, VAR.existing.df, directory = "Lookups\\\\",
  by = NULL, ReplaceNAsColumns = NULL, LookupTrumps = TRUE,
  NA.check.columns = NULL, OnlyKeepCheckedColumns = FALSE)
}
\arguments{
\item{VAR.path}{The location of the lookup file}

\item{VAR.file}{The name of the lookup file}

\item{VAR.existing.df}{The data frame to be joined}

\item{directory="Lookups\\"}{The directory within the path that holds the lookup}

\item{by=NULL}{The columns used to join, if not provided, matching columns will be used}

\item{ReplaceNAsColumns=NULL}{Before the join, these columns will have NAs values replaced}

\item{LookupTrumps=TRUE}{Should the function replace for common columns not used to join?}

\item{NA.check.columns=NULL, }{What new columns should be checked for NA values?}

\item{OnlyKeepCheckedColumns=FALSE}{Should only checked new columns be kept?}
}
\value{
The data frame plus new columns from the lookup file. If OnlyKeepCheckedColumns is
true and only new columns listed in NA.check.columns will be kept. Note to self, should
add input protection that throws an error if OnlyKeepCheckedColumns is set to true when
NA.check.columns is false.
}
\description{
Read in an external file and join it with an existing data frame.
}
\examples{
NA.check(VAR.existing.df,
  VAR.input=by,
  VAR.output=NA.check.columns,
  VAR.file=VAR.file)

}
