% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{read_and_join_experiment}
\alias{read_and_join_experiment}
\title{Read in an external file and join it with an existing data frame}
\usage{
read_and_join_experiment(
  data,
  lookup_file,
  path = "https://raw.githubusercontent.com/CSISdefense/R-scripts-and-data/master/",
  directory = "Lookups/",
  by = NULL,
  replace_na_var = NULL,
  overlap_var_replaced = TRUE,
  add_var = NULL,
  new_var_checked = TRUE,
  skip_check_var = NULL,
  zip_file = NULL,
  col_types = NULL,
  case_sensitive = TRUE,
  missing_file = NULL,
  create_lookup_rdata = FALSE,
  lookup_char_as_factor = FALSE,
  guess_max = NULL
)
}
\arguments{
\item{data}{The data frame to be joined}

\item{lookup_file}{The name of the lookup file}

\item{path}{The location of the lookup file}

\item{directory}{The directory within the path that holds the lookup}

\item{by}{The columns used to join, if not provided, matching columns will be used}

\item{replace_na_var}{Before the join, these columns will have NAs values replaced}

\item{overlap_var_replaced}{Should the function replace for common columns not used to join?}

\item{add_var}{What new columns should be checked for NA values?}

\item{new_var_checked}{Should only checked new columns be kept?}

\item{skip_check_var}{List of vars that should not be checked for NA values}

\item{zip_file}{The source zip file.}

\item{missing_file}{Filename to output any unmatched variables for easy of processing}

\item{create_lookup_rdata}{Whether to create a rdata file using the lookup for ease of future inputing}
}
\value{
The data frame plus new columns from the lookup file. If new_var_checked is
true and only new columns listed in add_var will be kept. Note to self, should
add input protection that throws an error if new_var_checked is set to true when
add_var is false.
}
\description{
Read in an external file and join it with an existing data frame
}
\details{
This function is an elaborate join with various quality check measures thrown in.
At its simplest, it just joins the existing data frame with the passed file. But along the way
it will make some fixes to common CSV errors and also take advantage of some known facts about
how CSIS data is organized.
}
\examples{
na_check(data,
  input_var=by,
  output_var=add_var,
  lookup_file=lookup_file)

}
