% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{remove_bom}
\alias{remove_bom}
\title{Remove byte order marks, these appear before the first column name:}
\usage{
remove_bom(data)
}
\arguments{
\item{data}{A /code(data.frame)}
}
\value{
/code(data) with the first column stripped of a byte order mark, if present
}
\description{
Remove byte order marks, these appear before the first column name:
}
\details{
Byte order marks are included at the start of UTF
text files and if not properly processed on read in, can add
nonsense characters to the start of the first column of a the
/code(data.frame) read in from a text file.
}
\examples{
remove_bom(data)

}
