% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{remove_unlabeled}
\alias{remove_unlabeled}
\title{Remove unlabeled data, if it's less than 1 percent of total}
\usage{
remove_unlabeled(df, value_var, var_list, threshold = 0.99)
}
\arguments{
\item{df}{A data frame}

\item{threshold}{Percent of data that must be labeled, defaults 0.99}

\item{col_list}{list of columns to strip out unlabeled}
}
\value{
df
}
\description{
Remove unlabeled data, if it's less than 1 percent of total
}
