% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{rename_value}
\alias{rename_value}
\title{Renames a factor level to user-specified name, in the passed data frame}
\usage{
rename_value(data, input)
}
\arguments{
\item{data}{The data frame in which to rename the value}

\item{input}{Shiny input object}
}
\value{
A data frame with the factor level renamed
}
\description{
Renames a factor level to user-specified name, in the passed data frame
}
