% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLookups.R
\name{replace_nas_with_unlabeled}
\alias{replace_nas_with_unlabeled}
\title{Replace NAs in one column of a data frame with a specified valued}
\usage{
replace_nas_with_unlabeled(VAR.df, VAR.column, replacement = "Unlabeled")
}
\arguments{
\item{VAR.df}{A data frame}

\item{VAR.column}{The column to have NAs removed}

\item{replacement}{Charater string that will replace NAs, by default "Unlabeled"}
}
\value{
Returns a data frame with the VAR.column values replaced with replacement,
otherwise identical to the original data frame
}
\description{
Replace NAs in one column of a data frame with a specified valued
}
\examples{
VAR.long.DF<-replace_nas_with_unlabeled(VAR.df=data.DF
  ,VAR.column="SubCustomer"),
  replacement = "Uncategorized",
  )

}
