% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize_variable_names}
\alias{standardize_variable_names}
\title{Standardize variable names}
\usage{
standardize_variable_names(VAR.Path, VAR.df)
}
\arguments{
\item{VAR.path}{The location of the lookup file}

\item{VAR.existing.df}{The data frame to be joined}
}
\value{
VAR.existing.df with standardized column names.
}
\description{
Standardize variable names
}
\examples{
FullData<-standardize_variable_names(Path,
  FullData)

}
