% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{swap_in_zip}
\alias{swap_in_zip}
\title{When passed a file that doesn't exist, check for the zip version}
\usage{
swap_in_zip(filename, path, directory)
}
\arguments{
\item{filename}{The name of the data file}

\item{path=""}{The location of the data file#'}

\item{directory="Data\\"}{The directory within the path that holds the lookup}
}
\value{
The filename name, if the file exists. Otherwise the zip file that contains the file.
}
\description{
When passed a file that doesn't exist, check for the zip version
}
\details{
This function is meant for large data files. Our default approach is to zip them
up, one zip per file, and name the zip file the same name as the data file, except with zip
as an extension instead of .txt or .csv. This checks if the base file is available and if
not it handles opening the zip file instead.
}
