% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{text_to_bit}
\alias{text_to_bit}
\title{Get Column Key based on the names in a data frame}
\usage{
text_to_bit(x)
}
\arguments{
\item{x}{Text / numerical expression of a bit}
}
\value{
A 0,1,NULL value.
}
\description{
Get Column Key based on the names in a data frame
}
\details{
Converts from character/number to bit
}
\examples{

text_to_bit(c('Y','1','$20.00'))

}
