% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{text_to_number}
\alias{text_to_number}
\title{Get Column Key based on the names in a data frame}
\usage{
text_to_number(x)
}
\arguments{
\item{x}{A list of numbers stored as characters or factor}
}
\value{
The list in numerical format
}
\description{
Get Column Key based on the names in a data frame
}
\details{
Converts from factor/character to number and strips
out commas and $s along the way.
}
\examples{

text_to_number(c('10','1,000','$20.00'))

}
