% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{update_sample_col_CSIScontractID}
\alias{update_sample_col_CSIScontractID}
\title{Update a sample using a larger data frame.}
\usage{
update_sample_col_CSIScontractID(
  smp,
  full,
  col = NULL,
  drop_and_replace = FALSE
)
}
\arguments{
\item{smp}{A data frame of contracts ready for statistical analysis, which must contain CSIScontractID.}

\item{full}{A data frame of contracts with no key missing data and which must contain CSIScontractID.}

\item{col}{Speific columns to add, if blank, add all in full missing from sample}

\item{drop_and_replace}{If true, drop rows from sample missing from full. Then replace them with new rows from full.}
}
\value{
The updated sample
}
\description{
Update a sample using a larger data frame.
}
\details{
This is a function that updates samples using an updated
version of the population, e.g. new columns, and adds them to
existing samples. This might be used if a new column has been added
from SQL or if NA values are found in a oolumn being used in.
This isn't appropriate if the larger being drawn from has changed
in make up, for example adding a new years data.
}
\examples{
update_sample_col_CSIScontractID(smp,def[complete,],drop_and_replace=TRUE)

}
