% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPDS_breakdowns_functions.R
\name{update_title}
\alias{update_title}
\title{Populates the title field with a dynamic title, if appropriate}
\usage{
update_title(passed_data, input, user_title,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{passed_data}{The data used in the plot}

\item{input}{Shiny input object}

\item{user_title}{"None" unless the user has manually entered a title}

\item{session=getDefaultReactiveDomain()}{The shiny app session}
}
\value{
Returns???
}
\description{
Populates the title field with a dynamic title, if appropriate
}
