% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{write_twice}
\alias{write_twice}
\title{Save a csv in two locations}
\usage{
write_twice(data, first_path, second_path, dir, file)
}
\arguments{
\item{first_path}{the primary storage path}

\item{second_path}{the second path, typically sharepoint}

\item{dir}{what directory for the output}

\item{file}{filename}

\item{df}{a ggplot object}
}
\value{
no value
}
\description{
Save a csv in two locations
}
