% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{add_preassigned_scales}
\alias{add_preassigned_scales}
\title{Take existing data frame and associate colors with values}
\usage{
add_preassigned_scales(plot, labels_and_colors, var = "None")
}
\arguments{
\item{plot}{The existing ggplot, needed to add more than one scale}

\item{labels_and_colors}{A csis360 lookup data.frame with factor information}

\item{var}{The name of the column, default to none for generic colors}
}
\value{
A plot with added color and fill scales for the column passed
}
\description{
Take existing data frame and associate colors with values
}
\details{
labels_and_colors is a data.frame produced by
the csis360 package, see prepare_labels_and_colors drawing
from lookup table that preassign labels, order, and color
to the expected values of a column.
}
\examples{
plot<-get_preasssigned_scales(plot,labels_and_colors,"pricing.mechanism.sum")

}
