% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_Lookups.R
\name{apply_standard_lookups}
\alias{apply_standard_lookups}
\title{Add column variants, including summaries, and deflated dollars, to a contract dataset.}
\usage{
apply_standard_lookups(
  df,
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/"
)
}
\arguments{
\item{df}{A data frame.}

\item{path}{The path or url for the column key.  By default, checks
the CSISdefense Github lookups repository at CSISdefense/Lookup/master/}
}
\value{
The fiscal year of the date.
}
\description{
Add column variants, including summaries, and deflated dollars, to a contract dataset.
}
\details{
Add column variants, including summaries, and deflated dollars,
to a contract dataset.
}
\examples{


}
