% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{build_plot}
\alias{build_plot}
\title{Adds a geom layer to a ggplot object based on user input.}
\usage{
build_plot(data, chart_geom = "Line Chart", share = FALSE,
  x_var = NULL, y_var, color_var = "None", facet_var = "None",
  second_var = NULL, legend = TRUE, caption = TRUE,
  labels_and_colors = NULL, column_key = NULL, format = FALSE)
}
\arguments{
\item{data}{A tibble of formatted data for the ggplot}

\item{chart_geom}{choose from Line Chart or Bar Chart; defaults to Line Chart}

\item{share}{If TRUE, calculates the share}

\item{x_var}{The name of fiscal year variable, as string}

\item{y_var}{The name of variable to plot on y-axis}

\item{color_var}{The name of the coloration variable, as string}

\item{facet_var}{The name of facet variable, as string}

\item{second_var}{The name of the secondary facet variable, as string}

\item{legend}{If TRUE, includes a legend}

\item{caption}{If TRUE, includes a source caption}

\item{labels_and_colors}{A csis360 lookup data.frame with factor information}

\item{column_key}{A csis360 lookup data.frame with column information}
}
\value{
A ggplot object including user-specified geom layer
}
\description{
Adds a geom layer to a ggplot object based on user input.
}
\details{
Intended to handle ggplot settings that depend on user input.
Settings that apply universally should be added in server.R.
}
\examples{

}
