% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Labels.R
\name{date_x_year_breaks}
\alias{date_x_year_breaks}
\title{Quickly assign yearly breaks to a chart}
\usage{
date_x_year_breaks(
  start,
  stop,
  by,
  fiscal_year = TRUE,
  partial_year = NULL,
  partial_label = "\\nYTD"
)
}
\arguments{
\item{start}{First year of break sequence}

\item{stop}{Last year of break sequence}
}
\value{
A plot with added color and fill scales for the column passed
}
\description{
Quickly assign yearly breaks to a chart
}
\details{
Add year breaks at specified intervals for date data
}
\examples{
date_x_year_breaks(2000,2023,2)

}
