% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{get_plot_theme}
\alias{get_plot_theme}
\title{Get Plot Theme}
\usage{
get_plot_theme(erase_legend_title = TRUE, blank_x_lines = TRUE)
}
\value{
Thematic elements that can be added to a plot.
}
\description{
Get Plot Theme
}
\details{
A standard set of style elements that is applied
across csis360 plots by default. These can be overriden,
by adding different theme elements after adding this to
your plot.
}
\examples{
FullData<-standardize_variable_names(Path,
  FullData)

}
