% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{group_data_for_plot}
\alias{group_data_for_plot}
\title{Returns data in the appropriate format for the user-specified plot}
\usage{
group_data_for_plot(data, x_var, y_var, breakout, aggregate = "sum")
}
\arguments{
\item{data}{The data to format for the plot, as a tibble}

\item{x_var}{x-axis}

\item{y_var}{y-axis}

\item{breakout}{the facets or other divisions that will be grouped by when summing}

\item{aggregate}{aggregation function; defaults to sum}
}
\value{
A tibble of formatted data
}
\description{
Returns data in the appropriate format for the user-specified plot
}
\details{

}
