% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Style.R
\name{jitter_binary}
\alias{jitter_binary}
\title{Create a function to jitter the binary outcome (p89)}
\usage{
jitter_binary(a, jitt = 0.05)
}
\arguments{
\item{a}{???????????????????}

\item{jitt}{????????should jitt always equal .05 or does this number just have to be between 0 and 1, and .05 is the default?}
}
\value{
Takes a binary variable and randomly spreads it within 0 and 1
}
\description{
Create a function to jitter the binary outcome (p89)
}
\details{
From Andrew Gelman and Jennifer Hill. For use in graphs
to show the quantity of variable that otherwise would be stuck
on a single lines at 0 and 1. This is different from geom_jitter
because it keeps all the data within the possible range of the variable,
not allowing values greater than 1 or less than 0.
}
\examples{
jitter_binary(contract$b_CBre)

}
