% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{LimitData}
\alias{LimitData}
\title{dfghjk}
\usage{
LimitData(Path, FPDS.gov, customer = NULL, big.ProdServ = NULL)
}
\arguments{
\item{Path}{d}

\item{FPDS.gov}{d}

\item{customer}{d}

\item{big.ProdServ}{d}
}
\value{

}
\description{
dfghjk
}
\details{
Quality control function
}
