% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Footing.R
\name{load.FPDS.gov.customers.df}
\alias{load.FPDS.gov.customers.df}
\title{Returns a dataframe with customer data for quality control check}
\usage{

  load.FPDS.gov.customers.df(Path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/",
  prod_serv_blank = "", customer = "")
}
\arguments{
\item{Path}{file path name where Lookup files are stored}

\item{prod_serv_blank}{character string; choose from "Products", "Services", or ""; defaults to "" for all}

\item{customer}{character string; choose from "Defense", "DHS", "State and IAP", or ""; defaults to "" for all}
}
\value{
dataframe
}
\description{
Returns a dataframe with customer data for quality control check
}
\details{
Quality control function
}
