% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize_variable_names}
\alias{standardize_variable_names}
\title{Standardize variable names}
\usage{
standardize_variable_names(
  data,
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/style/",
  var = NULL,
  replace_special = FALSE
)
}
\arguments{
\item{data}{the data frame to be joined}

\item{path}{the location of the lookup file}

\item{var}{the variable names to standardize; by default all will be done}

\item{replace_special}{whether to replaces spaces and special characters in column names with periods}
}
\value{
Data with standardized variable names.
}
\description{
Standardize variable names
}
\details{
This function is designed to prepare CSIS data files for lookup
application. It primarily smooths out variation between different ways we've
written SQL statements. It relies on a pre-existing table of variant variable names.
The variable names are matched against that table in a case insensitive manner,
though no other procedural standardization is applied at this time.
}
\examples{
FullData<-standardize_variable_names(
  FullData,
  Path)

}
