?% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize_variable_names}
\alias{standardize_variable_names}
\title{Standardize variable names}
\usage{
standardize_variable_names(data,
  path = "https://raw.githubusercontent.com/CSISdefense/Lookup-Tables/master/style/",
  var = NULL)
}
\arguments{
\item{data}{The data frame to be joined}

\item{path}{The location of the lookup file}

\item{var}{The names to standard, by default all will be done}
}
\value{
data with standardized var names.
}
\description{
Standardize variable names
}
\examples{
FullData<-standardize_variable_names(
  FullData,
  Path)

}
